+++ src/Rib/App.hs
@@ -97,6 +97,7 @@ runWith src dst buildAction = \case
           shakeOptions
             { shakeVerbosity = Chatty,
               shakeRebuild = bool [] [(RebuildNow, "**")] fullGen,
+              shakeLintInside = [toFilePath src, toFilePath dst],
               shakeExtra = addShakeExtra (Dirs (src, dst)) (shakeExtra shakeOptions)
             }
      in shakeForward opts buildAction